tools/model_explorer_circle/pyproject.toml
[project]
name = "model-explore-circle"
version = "0.0.1"

dependencies = [
    "ai-edge-model-explorer",
]
